+++ database/migrations/2025_02_05_143034_rating.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('rating', function (Blueprint $table) {
             $table->id();
             $table->integer('star');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('product_id');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
